requirements.txt
ansible==11.3.0
ansible-core==2.18.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
docker==7.1.0
Flask==3.1.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-compute==1.28.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.2
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
resolvelib==1.0.1
rsa==4.9
urllib3==2.3.0
Werkzeug==3.1.3
